refactor: move computePlayCounts and computeGenreCounts to lib/stats
Pure data transforms extracted from api/spotify.ts into a new lib/stats.ts
so the api layer only handles network calls. useSpotifyAuth now imports them
from lib instead of api.
