+++ client/src/lib/index.ts
@@ -4,4 +4,5 @@ export * from './format';
 export * from './heatmap';
 export * from './marathons';
 export * from './obsession';
+export * from './stats';
 export * from './theme';
